Adds a conditional action to build and push rpm and apt artifacts for pull requests to the dev repositories in nexus. (#84)
Adds build action to publish artifacts to the dev repo. Also changed nightly versions to be published to the dev repo instead of the nightly repo.

For PRs the version number is changed before building by adding the pr number as a suffix.

Suppress publishing of artifacts for runs that are triggered externally (dependabot mostly) as these do not have access to the necessary secret for deploying artifacts.